+++ graylog2-web-interface/src/components/scratchpad/Scratchpad.tsx
