+++ graylog2-web-interface/src/routing/App.tsx
@@ -21,7 +21,7 @@ import chroma from 'chroma-js';
 
 import { ScratchpadProvider } from 'contexts/ScratchpadProvider';
 import { Icon, Spinner } from 'components/common';
-import Scratchpad from 'components/scratchpad';
+import Scratchpad from 'components/scratchpad/Scratchpad';
 import CurrentUserContext from 'contexts/CurrentUserContext';
 import Navigation from 'components/navigation/Navigation';
 import ReportedErrorBoundary from 'components/errors/ReportedErrorBoundary';
